+++ pkg/providers/edgeworkers/testdata/TestDataEdgeWorkersEdgeWorker/edgeworker_no_versions.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_edgeworker" "test" {
+  edgeworker_id = 5
+  local_bundle  = "test_tmp/TestDataEdgeWorkersEdgeWorker/bundles/no_versions.tgz"
+}