+++ pkg/providers/edgeworkers/testdata/TestDataEdgeWorkersEdgeWorker/edgeworker_one_warning.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_edgeworker" "test" {
+  edgeworker_id = 3
+  local_bundle  = "test_tmp/TestDataEdgeWorkersEdgeWorker/bundles/edgeworker_one_warning.tgz"
+}